pulumi/app/src/requirements.txt
google-cloud-translate>=3.0.0,<4.0.0
google-auth>=2.0.0,<3.0.0
